Update refactor pubusb + add utisl
