+++ pubsub/utils.go
@@ -1,0 +1,69 @@
+package pubsub
+
+import (
+	"context"
+	"crypto/x509"
+	"encoding/json"
+	"fmt"
+	"strings"
+)
+
+// Fetches system certs and returns them if possible. If unable to fetch system certs then an empty cert pool is returned instead.
+func getCerts() *x509.CertPool {
+	if certs, err := x509.SystemCertPool(); err == nil {
+		return certs
+	}
+
+	return x509.NewCertPool()
+}
+
+// parseUnionFields parses the schema JSON to identify avro union fields.
+func parseUnionFields(_ context.Context, schemaJSON string) (map[string]struct{}, error) {
+	var schema map[string]interface{}
+	if err := json.Unmarshal([]byte(schemaJSON), &schema); err != nil {
+		return nil, fmt.Errorf("failed to parse schema: %w", err)
+	}
+
+	unionFields := make(map[string]struct{})
+	fields := schema["fields"].([]interface{})
+	for _, field := range fields {
+		f := field.(map[string]interface{})
+		fieldType := f["type"]
+		if types, ok := fieldType.([]interface{}); ok && len(types) > 1 {
+			unionFields[f["name"].(string)] = struct{}{}
+		}
+	}
+	return unionFields, nil
+}
+
+// flattenUnionFields flattens union fields decoded from Avro.
+func flattenUnionFields(_ context.Context, data map[string]interface{}, unionFields map[string]struct{}) map[string]interface{} {
+	flatData := make(map[string]interface{})
+	for key, value := range data {
+		if _, ok := unionFields[key]; ok { // Check if this field is a union
+			if valueMap, ok := value.(map[string]interface{}); ok && len(valueMap) == 1 {
+				for _, actualValue := range valueMap {
+					flatData[key] = actualValue
+					break
+				}
+			} else {
+				flatData[key] = value
+			}
+		} else {
+			flatData[key] = value
+		}
+	}
+
+	return flatData
+}
+
+// checks connection error.
+func connErr(err error) bool {
+	msg := err.Error()
+
+	return strings.Contains(msg, "is unavailable")
+}
+
+func invalidReplayIDErr(err error) bool {
+	return strings.Contains(strings.ToLower(err.Error()), "replay id validation failed")
+}
